requirements.txt
apscheduler==3.5.1
argparse==1.4.0
gitpython==2.1.11
PyYAML==3.13
